requirements.txt
Django==1.6.5
South==0.8.4
argparse==1.2.1
dj-database-url==0.3.0
factory-boy==2.3.1
ipdb==0.8
ipython==2.1.0
psycopg2==2.5.3
wsgiref==0.1.2
Django==1.6.5
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
dj-database-url==0.3.0
django-extensions==1.3.7
django-grappelli==2.5.3
factory-boy==1.1.3
ipdb==0.8
ipython==2.1.0
psycopg2==2.5.3
six==1.6.1
wsgiref==0.1.2
